♻️ Refactor the localizations fetcher
